refactor: move pages typescript files to separate folders, rename personalCard to detailsCard, productCard to detailsCard, move sortAndSearch to Home folder, make variables, functions and files start from small letter
